Feature : [#3] move admin test
